+++ apps/designer-standalone/src/app/AzureLogicAppsDesigner/laDesigner.tsx
@@ -3,7 +3,6 @@ import { DesignerCommandBar } from './DesignerCommandBar';
 import type { ConnectionAndAppSetting, ConnectionsData, ParametersData } from './Models/Workflow';
 import { Artifact } from './Models/Workflow';
 import type { WorkflowApp } from './Models/WorkflowApp';
-import { ApiManagementService } from './Services/ApiManagement';
 import { ArtifactService } from './Services/Artifact';
 import { ChildWorkflowService } from './Services/ChildWorkflow';
 import { FileSystemConnectionCreationClient } from './Services/FileSystemConnectionCreationClient';
@@ -22,7 +21,7 @@ import { ArmParser } from './Utilities/ArmParser';
 import { WorkflowUtility } from './Utilities/Workflow';
 import { Chatbot } from '@microsoft/chatbot';
 import {
-  ApiManagementInstanceService,
+  BaseApiManagementService,
   BaseAppServiceService,
   BaseFunctionService,
   BaseGatewayService,
@@ -39,6 +38,8 @@ import { clone, equals, guid, isArmResourceId } from '@microsoft/utils-logic-app
 import type { LogicAppsV2 } from '@microsoft/utils-logic-apps';
 import isEqual from 'lodash.isequal';
 import * as React from 'react';
+import type { QueryClient } from 'react-query';
+import { useQueryClient } from 'react-query';
 import { useSelector } from 'react-redux';
 
 const apiVersion = '2020-06-01';
@@ -71,6 +72,7 @@ const DesignerEditor = () => {
   const connectionsData = data?.properties.files[Artifact.ConnectionsFile] ?? {};
   const connectionReferences = WorkflowUtility.convertConnectionsDataToReferences(connectionsData);
   const parameters = data?.properties.files[Artifact.ParametersFile] ?? {};
+  const queryClient = useQueryClient();
 
   const onRunInstanceSuccess = async (runDefinition: LogicAppsV2.RunInstanceDefinition) => {
     if (monitoringView) {
@@ -128,7 +130,8 @@ const DesignerEditor = () => {
         addConnectionData,
         getConnectionConfiguration,
         tenantId,
-        canonicalLocation
+        canonicalLocation,
+        queryClient
       ),
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [workflow, workflowId, connectionsData, settingsData, workflowAppData, tenantId, designerID]
@@ -234,7 +237,8 @@ const getDesignerServices = (
   addConnection: (data: ConnectionAndAppSetting) => Promise<void>,
   getConfiguration: (connectionId: string) => Promise<any>,
   tenantId: string | undefined,
-  location: string
+  location: string,
+  queryClient: QueryClient
 ): any => {
   const siteResourceId = new ArmParser(workflowId).topmostResourceId;
   const armUrl = 'https://management.azure.com';
@@ -269,11 +273,12 @@ const getDesignerServices = (
       }),
     },
   });
-  const apimService = new ApiManagementInstanceService({
+  const apiManagementService = new BaseApiManagementService({
     apiVersion: '2019-12-01',
     baseUrl,
     subscriptionId,
     httpClient,
+    queryClient,
   });
   const childWorkflowService = new ChildWorkflowService({ apiVersion, baseUrl: armUrl, siteResourceId, httpClient, workflowName });
   const artifactService = new ArtifactService({
@@ -283,7 +288,6 @@ const getDesignerServices = (
     httpClient,
     integrationAccountCallbackUrl: undefined,
   });
-  const apiManagementService = new ApiManagementService({ service: apimService });
   const appService = new BaseAppServiceService({ baseUrl: armUrl, apiVersion, subscriptionId, httpClient });
   const connectorService = new StandardConnectorService({
     apiVersion,
@@ -316,10 +320,8 @@ const getDesignerServices = (
         return apiManagementService.getOperationSchema(configuration.connection.apiId, parameters.operationId, isInput);
       },
       getSwaggerOperationSchema: (args: any) => {
-        const { parameters, nodeMetadata, isInput } = args;
-        const swaggerUrl = nodeMetadata?.['apiDefinitionUrl'];
-        if (!swaggerUrl) return Promise.resolve();
-        return appService.getOperationSchema(swaggerUrl, parameters.operationId, isInput);
+        const { parameters, isInput } = args;
+        return appService.getOperationSchema(parameters.swaggerUrl, parameters.operationId, isInput);
       },
     },
     valuesClient: {
@@ -329,9 +331,8 @@ const getDesignerServices = (
         return artifactService.getMapArtifacts(mapType, mapSource);
       },
       getSwaggerOperations: (args: any) => {
-        const { nodeMetadata } = args;
-        const swaggerUrl = nodeMetadata?.['apiDefinitionUrl'];
-        return appService.getOperations(swaggerUrl);
+        const { parameters } = args;
+        return appService.getOperations(parameters.swaggerUrl);
       },
       getSchemaArtifacts: (args: any) => artifactService.getSchemaArtifacts(args.parameters.schemaSource),
       getApimOperations: (args: any) => {
@@ -396,12 +397,6 @@ const getDesignerServices = (
     httpClient,
   });
 
-  // const loggerService = new Stan({
-  //   resourceID: workflowId,
-  //   designerVersion: packagejson.dependencies['@microsoft/logic-apps-designer'],
-  //   designerID,
-  // });
-
   const runService = new StandardRunService({
     apiVersion,
     baseUrl,
@@ -419,7 +414,7 @@ const getDesignerServices = (
     loggerService: null,
     oAuthService,
     workflowService,
-    apimService,
+    apimService: apiManagementService,
     functionService,
     runService,
   };