+++ libs/designer-ui/src/lib/arrayeditor/index.tsx
@@ -63,6 +63,7 @@ export const ArrayEditor: React.FC<ArrayEditorProps> = ({
   labelProps,
   itemSchema,
   placeholder,
+  dataAutomationId,
   getTokenPicker,
   onChange,
   castParameter,
@@ -137,7 +138,7 @@ export const ArrayEditor: React.FC<ArrayEditorProps> = ({
   });
 
   return (
-    <div className="msla-array-editor-container">
+    <div className="msla-array-editor-container" data-automation-id={dataAutomationId}>
       {collapsed ? (
         <CollapsedArray
           labelProps={labelProps}