+++ libs/services/designer-client-services/src/lib/base/operations/index.ts
@@ -6,8 +6,3 @@ export * from './request';
 export * from './schedule';
 export * from './variables';
 export * from './operationgroups';
-export * from './inlineCode';
-export * from './flatFile';
-export * from './integrationAccount';
-export * from './liquid';
-export * from './xml';