+++ libs/services/designer-client-services/src/lib/consumption/operations/operationgroups.ts
@@ -1,75 +1,3 @@
-export const controlGroup = {
-  id: 'connectionProviders/control',
-  name: 'control',
-  properties: {
-    brandColor: '#8C3900',
-    description: 'Control operations',
-    displayName: 'Control',
-    capabilities: ['actions'],
-    iconUri: 'https://logicappsv2resources.blob.core.windows.net/icons/control.svg',
-  },
-};
-
-export const dateTimeGroup = {
-  id: 'connectionProviders/datetime',
-  name: 'datetime',
-  properties: {
-    brandColor: '#1F85FF',
-    description: 'Date Time operations',
-    displayName: 'Date Time',
-    capabilities: ['actions'],
-    iconUri: 'https://logicappsv2resources.blob.core.windows.net/icons/recurrence.svg',
-  },
-};
-
-export const httpGroup = {
-  id: 'connectionProviders/http',
-  name: 'http',
-  properties: {
-    brandColor: '#709727',
-    description: 'All Http operations',
-    displayName: 'HTTP',
-    capabilities: ['actions', 'triggers'],
-    iconUri: 'https://logicappsv2resources.blob.core.windows.net/icons/http.svg',
-  },
-};
-
-export const requestGroup = {
-  id: 'connectionProviders/request',
-  name: 'request',
-  properties: {
-    brandColor: '#009DA5',
-    description: 'Operations to handle inbound request to workflow and send workflow response',
-    displayName: 'Request',
-    iconUri: 'https://logicappsv2resources.blob.core.windows.net/icons/request.svg',
-    capabilities: ['actions', 'triggers'],
-  },
-};
-
-export const scheduleGroup = {
-  id: 'connectionProviders/schedule',
-  name: 'schedule',
-  properties: {
-    brandColor: '#1F85FF',
-    description: 'Schedule operations',
-    displayName: 'Schedule',
-    iconUri: 'https://logicappsv2resources.blob.core.windows.net/icons/recurrence.svg',
-    capabilities: ['actions', 'triggers'],
-  },
-};
-
-export const variableGroup = {
-  id: 'connectionProviders/variable',
-  name: 'variable',
-  properties: {
-    brandColor: '#770BD6',
-    description: 'All variable operations',
-    displayName: 'Variables',
-    iconUri: 'https://logicappsv2resources.blob.core.windows.net/icons/variable.svg',
-    capabilities: ['actions'],
-  },
-};
-
 export const inlineCodeGroup = {
   id: 'connectionProviders/inlineCode',
   name: 'inlineCode',
@@ -78,18 +6,6 @@ export const inlineCodeGroup = {
     description: 'Inline Code',
     displayName: 'Inline Code',
     iconUri: 'https://logicappsv2resources.blob.core.windows.net/icons/inline_code.svg',
-    capabilities: ['actions'],
-  },
-};
-
-export const dataOperationsGroup = {
-  id: 'connectionProviders/dataOperationNew',
-  name: 'dataOperationNew',
-  properties: {
-    brandColor: '#8c6cff',
-    description: 'Data Operations new',
-    displayName: 'Data Operations new',
-    iconUri: 'https://logicappsv2resources.blob.core.windows.net/icons/compose.svg',
     capabilities: ['actions'],
   },
 };
@@ -107,7 +23,7 @@ export const flatFileGroup = {
 };
 
 export const integrationAccountGroup = {
-  id: 'connectionProviders/integrationAccount',
+  id: 'connectionProviders/integrationAccountOperations',
   name: 'integrationAccount',
   properties: {
     brandColor: '#f1f1f1',